Factoring out server handling code into own class.
