+++ frontend/src/ts/widget.ts
@@ -1,7 +1,8 @@
 import {Area, OutputArea, LabContainer} from './areas';
 import {ButtonGroup, CheckBox, Tabs} from './components';
 import {Editor, EditorTheme} from './editor';
-import {fetchJSON, fetchBlob, DownloadRequest, DownloadResponse} from './comms';
+import {fetchBlob, DownloadRequest, DownloadResponse} from './comms';
+import {ServerWorker} from './server';
 import {Resource, RunProgram, CheckOutput} from './types';
 import * as Strings from './strings';
 import * as util from './utilities';
@@ -170,16 +171,13 @@ export class Widget {
       lab: lab,
     };
 
-    try {
-      const json =
-        await
-        fetchJSON<RunProgram.TS, RunProgram.FS>(serverData,
-            this.serverAddress('run_program'));
-      if (json.identifier == '') {
-        throw new Error(json.message);
-      }
+    const worker = new ServerWorker(this.server,
+        (data: CheckOutput.FS): number => {
+          return this.processCheckOutput(data);
+        });
 
-      await this.getOutputFromIdentifier(json);
+    try {
+      await worker.request(serverData, 'run_program');
     } catch (error) {
       this.outputArea.addError(Strings.MACHINE_BUSY_LABEL);
       console.error('Error:', error);
@@ -225,37 +223,6 @@ export class Widget {
     }
 
     return blobList;
-  }
-
-  /**
-   * Get the run output using the return identifier from the button CB
-   * @param {RunProgram.FS} json - the json data returned from button CB
-   * @param {number} lRead - the number of lines already read from the stream
-   * @param {number} nReq - the number of requests sent
-   */
-  private async getOutputFromIdentifier(json: RunProgram.FS,
-      lRead = 0, nReq = 0): Promise<void> {
-    const data: CheckOutput.TS = {
-      identifier: json.identifier,
-      read: lRead,
-    };
-
-    const rdata =
-      await fetchJSON<CheckOutput.TS, CheckOutput.FS>(data,
-          this.serverAddress('check_output'));
-    nReq++;
-
-    if (nReq >= 200) {
-      throw new Error('Request timed out. ' + Strings.INTERNAL_ERROR_MESSAGE);
-    }
-
-    lRead += this.processCheckOutput(rdata);
-
-    if (!rdata.completed) {
-      // We have not finished processing the output: call this again
-      await util.delay(250);
-      await this.getOutputFromIdentifier(json, lRead, nReq);
-    }
   }
 
   /**
